refactor: Update dependencies and imports in codebase
