+++ package.json
@@ -37,7 +37,7 @@
     "@typescript-eslint/eslint-plugin": "^4.28.4",
     "@typescript-eslint/parser": "^4.28.4",
     "aws-cli-js": "^2.2.2",
-    "aws-sdk": "^2.1643.0",
+    "aws-sdk": "^2.1691.0",
     "better-docs": "^2.7.2",
     "buffer": "^5.5.0||^6.0.0",
     "concurrently": "^7.1.0",
@@ -54,7 +54,7 @@
     "prettier": "^2.3.1",
     "start-server-and-test": "^1.14.0",
     "ts-node": "^10.8.1",
-    "typescript": "^4.3.5"
+    "typescript": "^5.6.3"
   },
   "browserslist": [
     "since 2017-06"
