+++ src/index.ts
@@ -1,4 +1,3 @@
-import loadAdapters from "src/adapters";
 import logger from "src/shared/logger";
 
 import { QueryStringContext } from "./shared/types";
@@ -16,16 +15,7 @@ import { getCustomTags } from "./shared/utils/get-custom-tags";
 
     getCustomTags();
 
-    await loadAdapters(context);
-
-    // switch (context.version) {
-    //   case "1":
-    //     import("./v1").then(({ default: load }) => load(context));
-    //     break;
-    //   case "2":
-    //     import("./v2").then(({ default: load }) => load(context));
-    //     break;
-    // }
+    await import("src/adapters").then(({ default: load }) => load(context));
   } catch (err) {
     const clientError = `An error has occured, please contact your pixel provider: `;
     console.error(clientError + err.message);
