+++ src/shared/extensions/bing-ads.ts
