+++ src/shared/extensions/segments.ts
